start mem complexity
