+++ pyccel/complexity/memory.py
@@ -10,30 +10,11 @@ the memory complexity a of a program.
 
 Example
 -------
-
->>> code = '''
-... n = 10
-... for i in range(0,n):
-...     for j in range(0,n):
-...         x = pow(i,2) + pow(i,3) + 3*i
-...         y = x / 3 + 2* x
-... '''
-
->>> from pyccel.complexity.memory import MemComplexity
->>> M = MemComplexity(code)
->>> d = M.cost()
->>> print "f = ", d['f']
-f =  n**2*(2*ADD + DIV + 2*MUL + 2*POW)
->>> print "m = ", d['m']
-m =  WRITE + 2*n**2*(READ + WRITE)
->>> q = M.intensity()
->>> print "+++ computational intensity ~", q
-+++ computational intensity ~ (2*ADD + DIV + 2*MUL + 2*POW)/(2*READ + 2*WRITE)
-
 """
 
 from sympy import sympify, Symbol
 from sympy import Poly, LT
+from sympy import summation
 
 from pyccel.ast.basic        import Basic
 from pyccel.ast.builtins     import PythonTuple
@@ -42,9 +23,14 @@ from pyccel.ast.internals    import PyccelSymbol
 from pyccel.ast.numpyext     import NumpyZeros, NumpyOnes
 from pyccel.ast.sympy_helper import pyccel_to_sympy
 from pyccel.complexity.basic import Complexity
+from pyccel.complexity.arithmetic import _compute_size_lhs
 
 
 __all__ = ["count_access", "MemComplexity"]
+
+WRITE = Symbol('WRITE')
+READ  = Symbol('READ')
+
 
 # ...
 def count_access(expr, visual=True):
@@ -62,25 +48,12 @@ def count_access(expr, visual=True):
         ignore their corresponding memory accesses.
     """
 
-    WRITE = Symbol('WRITE')
-    READ  = Symbol('READ')
 
     symbol_map = {}
     used_names = set()
 
     if isinstance(expr, Assign):
         return count_access(expr.rhs, visual) + WRITE
-
-    elif isinstance(expr, PythonTuple):
-        return sum(count_access(i, visual) for i in expr)
-
-    elif isinstance(expr, CodeBlock):
-        return sum(count_access(i, visual) for i in expr.body)
-
-    elif isinstance(expr, For):
-        s = pyccel_to_sympy(expr.iterable, symbol_map, used_names).size
-        ops = sum(count_access(i, visual) for i in expr.body.body)
-        return ops*s
 
     elif isinstance(expr, (NumpyZeros, NumpyOnes)):
         import numpy as np
@@ -94,23 +67,6 @@ def count_access(expr, visual=True):
     else:
         raise NotImplementedError('TODO count_access for {}'.format(type(expr)))
 
-
-
-def leading_term(expr, *args):
-    """
-    Returns the leading term in a sympy Polynomial.
-
-    expr: sympy.Expr
-        any sympy expression
-
-    args: list
-        list of input symbols for univariate/multivariate polynomials
-    """
-    expr = sympify(str(expr))
-    P = Poly(expr, *args)
-    return LT(P)
-# ...
-
 # ...
 class MemComplexity(Complexity):
     """
@@ -119,119 +75,87 @@ class MemComplexity(Complexity):
 
     Example
 
-    >>> code = '''
-    ... n = 10
-    ... for i in range(0,n):
-    ...     for j in range(0,n):
-    ...         x = pow(i,2) + pow(i,3) + 3*i
-    ...         y = x / 3 + 2* x
-    ... '''
-
-    >>> from pyccel.complexity.memory import MemComplexity
-    >>> M = MemComplexity(code)
-    >>> d = M.cost()
-    >>> print "f = ", d['f']
-    f =  n**2*(2*ADD + DIV + 2*MUL + 2*POW)
-    >>> print "m = ", d['m']
-    m =  WRITE + 2*n**2*(READ + WRITE)
-    >>> q = M.intensity()
-    >>> print "+++ computational intensity ~", q
-    +++ computational intensity ~ (2*ADD + DIV + 2*MUL + 2*POW)/(2*READ + 2*WRITE)
-
-    Now let us consider a case where some variables are supposed to be in the
-    fast memory, (*r* in this test)
-
-    >>> code = '''
-    ... n = 10
-    ... x = zeros(shape=(n,n), dtype=float)
-    ... r = float()
-    ... r = 0
-    ... for i in range(0, n):
-    ...     r = x[n,i] + 1
-    ... '''
-
-    >>> M = MemComplexity(code)
-    >>> d = M.cost()
-    >>> print "f = ", d['f']
-    f =  ADD*n
-    >>> print "m = ", d['m']
-    m =  2*WRITE + n*(READ + WRITE)
-    >>> q = M.intensity()
-    >>> print "+++ computational intensity ~", q
-    +++ computational intensity ~ ADD/(READ + WRITE)
-
-    Notice, that this is not what we expect! the cost of writing into *r* is
-    'zero', and therefor, there should be no :math:`n*WRITE` in our memory cost.
-    In order to achieve this, you must tell pyccel that you have the variable
-    *r* is already in the fast memory. This can be done by adding the argument
-    *local_vars=['r']* when calling the cost method.
-
-    >>> d = M.cost(local_vars=['r'])
-    >>> print "f = ", d['f']
-    f =  ADD*n
-    >>> print "m = ", d['m']
-    m =  READ*n + WRITE
-    >>> q = M.intensity(local_vars=['r'])
-    >>> print "+++ computational intensity ~", q
-    +++ computational intensity ~ ADD/READ
-
-    and this is exactly what we were expecting.
     """
-    def cost(self):
-        """
-        Computes the complexity of the given code.
 
-        local_vars: list
-            list of variables that are supposed to be in the fast memory. We will
-            ignore their corresponding memory accesses.
-        """
+    def _cost_Assign(self, expr, **settings):
+        ntimes = _compute_size_lhs(expr)
+        return ntimes * ( self._cost(expr.rhs, **settings) + WRITE )
 
-        return count_access(self.ast, visual=True)
+    def _cost_AugAssign(self, expr, **settings):
+        ntimes = _compute_size_lhs(expr)
+        return ntimes * ( self._cost(expr.rhs, **settings) + WRITE )
 
-    def intensity(self, d=None, args=None, local_vars=[], verbose=False):
-        """
-        Returns the computational intensity for the two level memory model.
+    def _cost_PyccelAdd(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        d: dict
-            dictionary containing the floating and memory costs. if not given,
-            we will compute them.
-        args: list
-            list of free parameters, i.e. degrees of freedom.
-        local_vars: list
-            list of variables that are supposed to be in the fast memory. We will
-            ignore their corresponding memory accesses.
-        verbose: bool
-            talk more
-        """
-        # ...
-        if d is None:
-            d = self.cost(local_vars=local_vars)
-        # ...
+    def _cost_PyccelMinus(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        # ...
-        if args is None:
-            args = self.free_parameters
-        # ...
+    def _cost_PyccelDiv(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        # ...
-        f = d['f']
-        m = d['m']
-        # ...
+    def _cost_PyccelFloorDiv(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        # ...
-        lt_f = leading_term(f, *args)
-        lt_m = leading_term(m, *args)
+    def _cost_PyccelMul(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        q = lt_f/lt_m
-        # ...
+    def _cost_PythonAbs(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
-        # ...
-        if verbose:
-            print((" arithmetic cost         ~ " + str(f)))
-            print((" memory cost             ~ " + str(m)))
-            print((" computational intensity ~ " + str(q)))
-        # ...
+    def _cost_Variable(self, expr, **settings):
+        return READ
 
-        return q
+    def _cost_PyccelSymbol(self, expr, **settings):
+        return READ
 
+    def _cost_NumpyFloor(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
 
+    def _cost_NumpyExp(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyLog(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpySqrt(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpySin(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyCos(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyTan(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArcsin(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArccos(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArctan(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArctan2(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpySinh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyCosh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyTanh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArcsinh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArccosh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
+
+    def _cost_NumpyArctanh(self, expr, **settings):
+        return sum(self._cost(i, **settings) for i in expr.args)
